Downloads/iran-war-notelm-urgentdash-live/iran-war-notelm-main/requirements.txt
playwright
apscheduler
python-telegram-bot
structlog
polars
python-dotenv
tenacity
notebooklm-mcp-cli
httpx
beautifulsoup4
pydantic-settings
